% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unbiased_mcmc.R
\name{beta_unbiased_histogram}
\alias{beta_unbiased_histogram}
\title{beta_unbiased_histogram}
\usage{
beta_unbiased_histogram(nchains, component, breaks = NULL, nclass = 30)
}
\arguments{
\item{nchains}{no of chains}

\item{component}{index of component}

\item{breaks}{histogram breaks}

\item{nclass}{histogram number of classes}
}
\value{
list containing histogram data with mids = mids, breaks = breaks, proportions = prop_mean, sd = prop_sd, width = width
}
\description{
Outputs unbiased estimates for a histogram of a component
}
