% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_half_t.R
\name{beta_update}
\alias{beta_update}
\title{beta_update}
\usage{
beta_update(xi, sigma2, eta, X, X_transpose, y, M_matrix_inverse, active_set)
}
\arguments{
\item{xi}{positive scalar}

\item{sigma2}{positive scalar}

\item{eta}{length p vector}

\item{X}{n by p matrix}

\item{X_transpose}{Pre-calculated transpose of X}

\item{y}{length n vector}

\item{M_matrix_inverse}{n by n matrix}

\item{active_set}{length p vector of booleans}
}
\value{
beta update given xi, eta, sigma2
}
\description{
beta given xi, eta, sigma2 update using algo of Bhattacharya
}
