% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eta_half_t_sampling.R
\name{conf_hyp}
\alias{conf_hyp}
\title{conf_hyp}
\usage{
conf_hyp(a, b, z)
}
\arguments{
\item{a}{positive scalar}

\item{b}{positive scalar}

\item{z}{positive scalar}
}
\value{
Returns \eqn{U(a,b,z) := \frac{1}{\Gamma(a)} \int_0^\infty x^{a-1}(1+x)^{b-a-1} e^{-zx}dx}
}
\description{
Inverse of Lower incomplete Gamma Function.
}
