% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coupled_mcmc_half_t.R
\name{coupled_half_t_kernel}
\alias{coupled_half_t_kernel}
\title{coupled_half_t_kernel}
\usage{
coupled_half_t_kernel(
  X,
  X_transpose,
  y,
  a0 = 1,
  b0 = 1,
  std_MH = 0.8,
  xi_1_current,
  xi_2_current,
  sigma2_1_current,
  sigma2_2_current,
  beta_1_current,
  beta_2_current,
  eta_1_current,
  eta_2_current,
  approximate_algo_delta = 0,
  epsilon_eta = 0.5,
  epsilon_xi = Inf,
  epsilon_sigma2 = Inf,
  nrepeats_eta = 1,
  verbose = FALSE,
  xi_fixed = FALSE,
  sigma2_fixed = FALSE,
  t_dist_df,
  two_scale = TRUE
)
}
\arguments{
\item{X}{n by p matrix}

\item{X_transpose}{Pre-calculated transpose of X}

\item{y}{length n vector}

\item{a0}{positive scalar}

\item{b0}{positive scalar}

\item{std_MH}{standard deviation of log-normal MH proposal}

\item{xi_1_current, xi_2_current}{current xi values (positive scalar)}

\item{sigma2_1_current, sigma2_2_current}{current sigma2 values (positive scalar)}

\item{beta_1_current, beta_2_current}{current beta values (vector of length p)}

\item{eta_1_current, eta_2_current}{current eta values (vector of length p)}

\item{approximate_algo_delta}{approximate MCMC error (non-negative scalar)}

\item{epsilon_eta}{eta common random numbers/ maximal coupling coupling threshold}

\item{epsilon_xi}{xi common random numbers/ maximal coupling coupling threshold}

\item{epsilon_sigma2}{sigma2 common random numbers/ maximal coupling coupling threshold}

\item{nrepeats_eta}{number of slice sampling steps}

\item{verbose}{boolean for printing/ not printing run time}

\item{xi_fixed}{boolean for fixing / not fixing xi}

\item{sigma2_fixed}{boolean for fixing / not fixing sigma2}

\item{t_dist_df}{degree of freedom v>=1 for Half-t(v).}
}
\value{
(beta1, beta2, eta1, eta2, sigma2_1, sigma2_2, xi1, xi2, metric_d)
}
\description{
Coupled blocked Gibbs kernel for half-t priors
}
