% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coupled_mcmc_half_t.R
\name{coupled_half_t_mcmc}
\alias{coupled_half_t_mcmc}
\title{coupled_half_t_mcmc}
\usage{
coupled_half_t_mcmc(
  mc_chain_size,
  X,
  X_transpose,
  y,
  a0 = 1,
  b0 = 1,
  std_MH = 0.8,
  rinit = NULL,
  approximate_algo_delta = 0,
  epsilon_eta = 0.5,
  epsilon_xi = Inf,
  epsilon_sigma2 = Inf,
  nrepeats_eta = 1,
  verbose = FALSE,
  preallocate = 100,
  max_iterations = Inf,
  totalduration = Inf,
  xi_fixed = FALSE,
  sigma2_fixed = FALSE,
  lag = 1,
  t_dist_df,
  two_scale = TRUE,
  xi_interval = c(0, Inf)
)
}
\arguments{
\item{mc_chain_size}{minimum length of Markov chain}

\item{X}{n by p matrix}

\item{X_transpose}{Pre-calculated transpose of X}

\item{y}{length n vector}

\item{a0}{positive scalar}

\item{b0}{positive scalar}

\item{rinit}{Initial distribution}

\item{approximate_algo_delta}{approximate MCMC error (non-negative scalar)}

\item{epsilon_eta}{eta common random numbers/ maximal coupling coupling threshold}

\item{epsilon_xi}{xi common random numbers/ maximal coupling coupling threshold}

\item{epsilon_sigma2}{sigma2 common random numbers/ maximal coupling coupling threshold}

\item{nrepeats_eta}{number of slice sampling steps}

\item{verbose}{boolean for printing/ not printing run time}

\item{preallocate}{pre-allocated memory}

\item{max_iterations}{number of maximum iterations}

\item{totalduration}{maximum duration}

\item{xi_fixed}{boolean for fixing / not fixing xi}

\item{sigma2_fixed}{boolean for fixing / not fixing sigma2}

\item{lag}{Lag}

\item{t_dist_df}{degree of freedom v>=1 for Half-t(v)}

\item{xi_interval}{support of prior distribution of xi}
}
\value{
(beta, eta, sigma2, xi) sampled from the half-t prior blocked Gibbs kernel
}
\description{
Coupled Blocked Gibbs MCMC for Bayesian shrinkage with half-t priors
}
