% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coupledHalfT-package.R
\docType{package}
\name{coupledHalfT-package}
\alias{coupledHalfT-package}
\alias{coupledHalfT}
\title{coupledHalfT}
\description{
Coupled Bayesian Shrinkage with Half-t(v) priors
}
\details{
This package contains scripts to reproduce the figures of the
paper "Coupled Bayesian Shrinkage with Half-t(v) priors"
}
\author{
Niloy Biswas
}
\keyword{package}
