% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coupled_mcmc_half_t.R
\name{crn_max_sigma2_coupling}
\alias{crn_max_sigma2_coupling}
\title{xi_update}
\usage{
crn_max_sigma2_coupling(
  xi_1,
  eta_1,
  xi_2,
  eta_2,
  n,
  ssr_1,
  ssr_2,
  a0,
  b0,
  epsilon_sigma2 = 0,
  sigma2_fixed_value = NULL
)
}
\arguments{
\item{xi_1, xi_2}{current xi values (positive scalars)}

\item{eta_1, eta_2}{current eta values (vector length p)}

\item{n}{number of data points}

\item{ssr_1, ssr_2}{postiive scalars}

\item{a0}{positive scalar}

\item{b0}{positive scalar}

\item{epsilon_sigma2}{threshold (when above/below use common random numbers/ maximal coupling of proposals)}

\item{sigma2_fixed_value}{boolean (when TRUE/FALSE sigma2 fixed/varying)}
}
\value{
Coupled update of sigma2 given eta
}
\description{
Coupled update of sigma2 given eta
}
