% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coupled_mcmc_half_t.R
\name{crn_max_xi_coupling}
\alias{crn_max_xi_coupling}
\title{xi_update}
\usage{
crn_max_xi_coupling(
  current_xi_1,
  eta_1,
  current_xi_2,
  eta_2,
  X,
  X_transpose,
  y,
  a0,
  b0,
  std_MH,
  approximate_algo_delta = 0,
  epsilon_xi = 0,
  fixed = FALSE,
  xi_interval
)
}
\arguments{
\item{current_xi_1, current_xi_2}{current xi values (positive scalars)}

\item{eta_1, eta_2}{current eta values (vector length p)}

\item{X}{n by p matrix}

\item{X_transpose}{Pre-calculated transpose of X}

\item{y}{length n vector}

\item{a0}{positive scalar}

\item{b0}{positive scalar}

\item{std_MH}{standard deviation of log-normal MH proposal}

\item{approximate_algo_delta}{approximate MCMC error (non-negative scalar)}

\item{epsilon_xi}{threshold (when above/below use common random numbers/ maximal coupling of proposals)}

\item{xi_interval}{support of prior distribution of xi}
}
\value{
Coupled Metropolis-Hastings update of xi given eta
}
\description{
Coupled Metropolis-Hastings update of xi given eta
}
