% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eta_half_t_sampling.R
\name{d_trunc_poly_exp}
\alias{d_trunc_poly_exp}
\title{d_trunc_poly_exp}
\usage{
d_trunc_poly_exp(poly_exponent, rate, trunc_upper, eta)
}
\arguments{
\item{poly_exponent}{positive scalar}

\item{rate}{positive vector}

\item{trunc_upper}{positive scalar}

\item{eta}{positive vector}
}
\value{
Returns Log Pdf of \eqn{p(x) \prop x^(poly_exponent-1)*exp(-rate*x) on [0,trunc_upper]}
}
\description{
Log Pdf of \eqn{ p(x) \prop x^(poly_exponent-1)*exp(-rate*x) on [0,trunc_upper] }
}
