% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eta_half_t_sampling.R
\name{eta_update_half_t}
\alias{eta_update_half_t}
\title{eta_update_half_t}
\usage{
eta_update_half_t(xi, sigma2, beta, eta, t_dist_df, nrepeats = 1)
}
\arguments{
\item{xi}{positive scalar}

\item{sigma2}{positive scalar}

\item{beta}{positive vector}

\item{eta}{positive vector}

\item{t_dist_df}{half-t degree of freedom}

\item{nrepeats}{number of slice sampling steps}
}
\value{
Slice sampler targeting \eqn{p(x) \prop x^((v-1)/2)/(1+vx)^((v+1)/2)*exp(-mx) on [0,Inf], }
where v is t_dist_df and m is rate.
}
\description{
Slice sampler targeting \eqn{p(x) \prop x^((v-1)/2)/(1+vx)^((v+1)/2)*exp(-mx) on [0,Inf], }
where v is t_dist_df and m is rate.
}
