% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_half_t.R
\name{half_t_kernel}
\alias{half_t_kernel}
\title{half_t_kernel}
\usage{
half_t_kernel(
  X,
  X_transpose,
  y,
  a0 = 1,
  b0 = 1,
  std_MH = 0.8,
  xi_current,
  sigma2_current,
  beta_current,
  eta_current,
  approximate_algo_delta,
  nrepeats_eta = 1,
  verbose = FALSE,
  xi_fixed = FALSE,
  sigma2_fixed = FALSE,
  t_dist_df,
  xi_interval
)
}
\arguments{
\item{X}{n by p matrix}

\item{X_transpose}{Pre-calculated transpose of X}

\item{y}{length n vector}

\item{a0}{positive scalar}

\item{b0}{positive scalar}

\item{std_MH}{standard deviation of log-normal MH proposal}

\item{xi_current}{current xi value (positive scalar)}

\item{sigma2_current}{current sigma2 value (positive scalar)}

\item{beta_current}{current beta value (vector of length p)}

\item{eta_current}{current eta value (vector of length p)}

\item{approximate_algo_delta}{approximate MCMC error (non-negative scalar)}

\item{nrepeats_eta}{number of slice sampling steps}

\item{verbose}{boolean for printing/ not printing run time}

\item{xi_fixed}{boolean for fixing / not fixing xi}

\item{sigma2_fixed}{boolean for fixing / not fixing sigma2}

\item{t_dist_df}{degree of freedom v for Half-t(v). Take v >=1.}

\item{xi_interval}{support of prior distribution of xi}
}
\value{
(beta, eta, sigma2, xi) sampled from the half-t prior blocked Gibbs kernel
}
\description{
Full blocked Gibbs kernel for half-t priors
}
