% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_half_t.R
\name{half_t_mcmc}
\alias{half_t_mcmc}
\title{half_t_mcmc}
\usage{
half_t_mcmc(
  chain_length,
  burnin,
  X,
  X_transpose,
  y,
  a0 = 1,
  b0 = 1,
  std_MH = 0.8,
  rinit = NULL,
  approximate_algo_delta = 0,
  nrepeats_eta = 1,
  verbose = FALSE,
  xi_fixed = FALSE,
  sigma2_fixed = FALSE,
  t_dist_df,
  xi_interval = c(0, Inf)
)
}
\arguments{
\item{chain_length}{length of Markov chain}

\item{burnin}{Markov chain burn in}

\item{X}{n by p matrix}

\item{X_transpose}{Pre-calculated transpose of X}

\item{y}{length n vector}

\item{a0}{positive scalar}

\item{b0}{positive scalar}

\item{rinit}{Initial distribution}

\item{approximate_algo_delta}{approximate MCMC error (non-negative scalar)}

\item{nrepeats_eta}{number of slice sampling steps}

\item{verbose}{boolean for printing/ not printing run time}

\item{xi_fixed}{boolean for fixing / not fixing xi}

\item{sigma2_fixed}{boolean for fixing / not fixing sigma2}

\item{t_dist_df}{degree of freedom v for Half-t(v). Take v >=1.}

\item{xi_interval}{support of prior distribution of xi}

\item{xi_current}{current xi value (positive scalar)}

\item{sigma2_current}{current sigma2 value (positive scalar)}

\item{beta_current}{current beta value (vector of length p)}

\item{eta_current}{current eta value (vector of length p)}
}
\value{
(beta, eta, sigma2, xi) sampled from the half-t prior blocked Gibbs kernel
}
\description{
Blocked Gibbs MCMC for Bayesian shrinkage with half-t priors
}
