% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_half_t.R
\name{log_ratio}
\alias{log_ratio}
\title{log_ratio}
\usage{
log_ratio(xi, eta, X_eta_tX_matrix, y, a0, b0)
}
\arguments{
\item{xi}{positive scalar}

\item{eta}{length p vector}

\item{X_eta_tX_matrix}{X * Diag(1/eta) * t(X_transpose), matrix n by n}

\item{y}{length n vector}

\item{a0}{positive scalar}

\item{b0}{positive scalar}
}
\value{
Unnormalized posterior pdf of log(xi)
}
\description{
Unnormalized posterior pdf of log(xi)
}
