% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_half_t.R
\name{log_ratio_approx}
\alias{log_ratio_approx}
\title{log_ratio_approx}
\usage{
log_ratio_approx(xi, eta, X, X_transpose, y, a0, b0, active_set, xi_interval)
}
\arguments{
\item{xi}{positive scalar}

\item{eta}{length p vector}

\item{X}{n by p matrix}

\item{X_transpose}{Pre-calculated transpose of X}

\item{y}{length n vector}

\item{a0}{positive scalar}

\item{b0}{positive scalar}

\item{active_set}{length p vector of booleans}

\item{xi_interval}{support of prior distribution of xi}
}
\value{
Unnormalized posterior pdf of log(xi)
}
\description{
Unnormalized posterior pdf of log(xi) for approximate MCMC
}
