% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eta_half_t_sampling.R
\name{low_inc_conf_hyp}
\alias{low_inc_conf_hyp}
\title{low_inc_conf_hyp}
\usage{
low_inc_conf_hyp(a, b, z, t, regularize = TRUE)
}
\arguments{
\item{a}{positive scalar}

\item{b}{positive scalar}

\item{z}{positive scalar}

\item{t}{positive scalar}
}
\value{
Returns \eqn{U(a,b,z,t) := \frac{1}{\Gamma(a)} \int_0^t x^{a-1}(1+x)^{b-a-1} e^{-zx}dx}
}
\description{
Lower Incomplete Confluent Hypergeometric Function of the second kind
}
