% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eta_half_t_sampling.R
\name{low_inc_gamma}
\alias{low_inc_gamma}
\title{low_inc_gamma}
\usage{
low_inc_gamma(rate, upper_truncation, log = TRUE)
}
\arguments{
\item{rate}{positive scalar}

\item{upper_truncation}{positive scalar}

\item{log}{boolean to return log of lower incomplete gamma}
}
\value{
log(pgamma(x,a)) or pgamma(x,a) where pgamma(x,a) := 1/Gamma(a) integral_0^x t^(a-1) exp(-t) dt.
}
\description{
Lower Incomplete Gamma Function.
}
