% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eta_half_t_sampling.R
\name{low_inc_gamma_inv}
\alias{low_inc_gamma_inv}
\title{low_inc_gamma_inv}
\usage{
low_inc_gamma_inv(rate, y, log = TRUE)
}
\arguments{
\item{rate}{positive scalar}

\item{y}{positive scalar}

\item{log}{boolean to return log of lower incomplete gamma}
}
\value{
Returns y s.t. low_inc_gamma(a, x)=y
}
\description{
Inverse of Lower incomplete Gamma Function.
}
\details{
qgamma is the quantile function of the gamma distribution
}
\seealso{
\code{\link{low_inc_gamma}} and \code{\link[zipfR]{Rgamma.inv()}}
}
