% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unbiased_mcmc.R
\name{plot_histogram}
\alias{plot_histogram}
\title{beta_unbiased_histogram}
\usage{
plot_histogram(histogram, with_bar = TRUE)
}
\arguments{
\item{with_bar}{boolean to include/exclude error bars}
}
\value{
histrogram plot
}
\description{
Plots histogram given histogram data
}
