% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eta_half_t_sampling.R
\name{r_trunc_poly_exp_crn}
\alias{r_trunc_poly_exp_crn}
\title{r_trunc_poly_exp_crn}
\usage{
r_trunc_poly_exp_crn(poly_exponent, rate, trunc_upper, unif)
}
\arguments{
\item{poly_exponent}{positive scalar}

\item{rate}{positive vector (function is vectorized)}

\item{trunc_upper}{positive scalar}

\item{unif}{positive scalar}
}
\value{
Returns vectorised samples from \eqn{p(x) \prop x^(poly_exponent-1)*exp(-rate*x) on [0,trunc_upper]}
}
\description{
Perfect sampling from \eqn{p(x) \prop x^(poly_exponent-1)*exp(-rate*x) on [0,trunc_upper]}
}
