% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unbiased_mcmc.R
\name{sample_unbiasedestimator}
\alias{sample_unbiasedestimator}
\title{sample_unbiasedestimator}
\usage{
sample_unbiasedestimator(
  single_kernel,
  coupled_kernel,
  rinit,
  h = function(x) x,
  k = 0,
  m = 1,
  lag = 1,
  max_iterations = Inf
)
}
\arguments{
\item{single_kernel}{kernel of marginal chain}

\item{coupled_kernel}{kernel of coupled chain}

\item{rinit}{initial distribution}

\item{h}{real-valued test function}

\item{k}{unbiased mcmc parameter corresponding to burn-in}

\item{m}{unbiased mcmc parameter corresponding to chain length}

\item{lag}{lag of the coupled chain}

\item{max_iterations}{maximum iterations}
}
\value{
list containing mcmc estimator, bias correction, unbiased estimator, meetinging, number of iterations, time elapsed, cost
}
\description{
Outputs unbiased estimates without storing the full chain
}
