% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_half_t.R
\name{X_eta_tX}
\alias{X_eta_tX}
\title{X_eta_tX}
\usage{
X_eta_tX(eta, X, X_transpose)
}
\arguments{
\item{eta}{vector of length p}

\item{X}{matrix of length n by p}

\item{X_transpose}{Pre-calculated transpose of X}
}
\value{
X Diag(1/eta) t(X_transpose)
}
\description{
Calculates X * Diag(1/eta) * t(X_transpose)
}
