% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_half_t.R
\name{xi_update}
\alias{xi_update}
\title{xi_update}
\usage{
xi_update(
  current_xi,
  eta,
  X,
  X_transpose,
  y,
  a0,
  b0,
  std_MH,
  approximate_algo_delta = 0,
  fixed = FALSE,
  xi_interval
)
}
\arguments{
\item{current_xi}{positive scalar}

\item{eta}{length p vector}

\item{X}{n by p matrix}

\item{X_transpose}{Pre-calculated transpose of X}

\item{y}{length n vector}

\item{a0}{positive scalar}

\item{b0}{positive scalar}

\item{std_MH}{standard deviation of log-normal MH proposal}

\item{approximate_algo_delta}{approximate MCMC error (non-negative scalar)}

\item{xi_interval}{support of prior distribution of xi}
}
\value{
Metropolis-Hastings update of xi given eta
}
\description{
Metropolis-Hastings update of xi
}
